+++ ShopifySharp/Entities/Customer.cs
@@ -110,9 +110,9 @@ namespace ShopifySharp
         /// <summary>
         /// The total amount of money that the customer has spent at the shop.
         /// </summary>
-        /// <remarks>The Shopify API actually returns this value as a string, but Json.Net can automatically convert to double.</remarks>
+        /// <remarks>The Shopify API actually returns this value as a string, but Json.Net can automatically convert to decimal.</remarks>
         [JsonProperty("total_spent")]
-        public double TotalSpent { get; set; }
+        public decimal TotalSpent { get; set; }
 
         /// <summary>
         /// The date and time when the customer information was updated.
