+++ ShopifySharp/Entities/ProductVariant.cs
@@ -60,13 +60,13 @@ namespace ShopifySharp
         /// The price of the product variant.
         /// </summary>
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         /// <summary>
         /// The competitors prices for the same item.
         /// </summary>
         [JsonProperty("compare_at_price", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public double? CompareAtPrice { get; set; }
+        public decimal? CompareAtPrice { get; set; }
 
         /// <summary>
         /// Custom properties that a shop owner can use to define product variants.
@@ -146,7 +146,7 @@ namespace ShopifySharp
         /// The weight of the product variant in the unit system specified with weight_unit.
         /// </summary>
         [JsonProperty("weight")]
-        public double Weight { get; set; }
+        public decimal Weight { get; set; }
 
         /// <summary>
         /// The unit system that the product variant's weight is measure in. The weight_unit can be either "g", "kg, "oz", or "lb".
