+++ ShopifySharp/Entities/TaxLine.cs
@@ -13,13 +13,13 @@ namespace ShopifySharp
         /// The amount of tax to be charged.    
         /// </summary>
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         /// <summary>
         /// The rate of tax to be applied.
         /// </summary>
         [JsonProperty("rate")]
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
 
         /// <summary>
         /// The name of the tax.
